move stream_results into inference, delete deprecated refact_scratchpads_no_gpu module
