+++ self_hosting_machinery/inference/stream_results_async.py
@@ -1,7 +1,14 @@
-import os, sys, json, re, time, datetime, termcolor, multiprocessing, copy, queue
+import os
+import json
+import time
+import datetime
+import termcolor
+import copy
 import aiohttp
 import asyncio
-from refact_scratchpads_no_gpu import stream_results
+
+from self_hosting_machinery.inference import stream_results
+
 from typing import Dict, Any, List, Optional, Set
